refactor: extract plugins.ts into plugin modules (WOP-17 Phase 3)
Break up the 1,332-line plugins.ts monolith into 12 focused modules
under src/plugins/:

- state.ts: Shared mutable state (Maps, constants)
- extensions.ts: Plugin extension registry
- schema-converter.ts: JSON Schema to Zod + A2A server registration
- installation.ts: Plugin install/remove/enable/disable
- context-factory.ts: WOPRPluginContext builder
- event-bus.ts: Plugin-scoped event bus factory
- hook-manager.ts: Plugin-scoped hook manager factory
- plugin-logger.ts: Plugin-scoped logger factory
- loading.ts: Plugin load/unload lifecycle
- registry.ts: Plugin discovery and registry management
- config-schemas.ts: Config schema accessors
- accessors.ts: Public read-only state accessors

plugins.ts is now a thin re-export barrel preserving the original
public API. All 113 tests pass, zero behavior changes.
