+++ src/plugins.ts
@@ -1,1332 +1,27 @@
-import { execFileSync, spawnSync } from "node:child_process";
-import { existsSync, mkdirSync, readFileSync, rmSync, symlinkSync, writeFileSync } from "node:fs";
-import { homedir } from "node:os";
-import { join, resolve } from "node:path";
-import { z } from "zod";
-import { registerA2ATool } from "./core/a2a-mcp.js";
-import {
-  getChannelProvider as getChannelProviderCore,
-  getChannelProviders as getChannelProvidersCore,
-  registerChannelProvider as registerChannelProviderCore,
-  unregisterChannelProvider as unregisterChannelProviderCore,
-} from "./core/channels.js";
-import { config as centralConfig } from "./core/config.js";
-import {
-  getContextProvider as getCtxProvider,
-  registerContextProvider as registerCtxProvider,
-  unregisterContextProvider as unregisterCtxProvider,
-} from "./core/context.js";
-import { eventBus } from "./core/events.js";
-import { providerRegistry } from "./core/providers.js";
-import { cancelInject as cancelSessionInject, logMessage as logMessageToSession } from "./core/sessions.js";
-import { resolveIdentity, resolveUserProfile } from "./core/workspace.js";
-import { logger } from "./logger.js";
-import { checkRequirements, ensureRequirements, formatMissingRequirements } from "./plugins/requirements.js";
-import type { ModelProvider } from "./types/provider.js";
-import type {
-  A2AServerConfig,
-  ChannelAdapter,
-  ChannelProvider,
-  ChannelRef,
-  ConfigSchema,
-  ContextProvider,
-  EventHandler,
-  HookOptions,
-  InstalledPlugin,
-  MutableHookEvent,
-  PluginInjectOptions,
-  PluginLogger,
-  PluginRegistryEntry,
-  UiComponentExtension,
-  WebUiExtension,
-  WOPREventBus,
-  WOPRHookManager,
-  WOPRPlugin,
-  WOPRPluginContext,
-} from "./types.js";
-import { getVoiceRegistry } from "./voice/index.js";
-import type { InstallMethod, STTProvider, TTSProvider, VoicePluginRequirements } from "./voice/types.js";
-
-const WOPR_HOME = process.env.WOPR_HOME || join(homedir(), "wopr");
-const PLUGINS_DIR = join(WOPR_HOME, "plugins");
-const PLUGINS_FILE = join(WOPR_HOME, "plugins.json");
-const REGISTRIES_FILE = join(WOPR_HOME, "plugin-registries.json");
-
-// Loaded plugins (runtime)
-const loadedPlugins: Map<string, { plugin: WOPRPlugin; context: WOPRPluginContext }> = new Map();
-
-// Context providers - session -> provider mapping
-const contextProviders: Map<string, ContextProvider> = new Map();
-const channelAdapters: Map<string, ChannelAdapter> = new Map();
-const webUiExtensions: Map<string, WebUiExtension> = new Map();
-const uiComponents: Map<string, UiComponentExtension> = new Map();
-
-// Provider plugins registry (for providers registered via plugins)
-const providerPlugins: Map<string, ModelProvider> = new Map();
-
-// Config schemas registry (pluginId -> schema)
-const configSchemas: Map<string, ConfigSchema> = new Map();
-
-// Plugin extensions registry - plugins can expose APIs to other plugins
-// Key format: "pluginName.extensionName" -> extension object
-const pluginExtensions: Map<string, unknown> = new Map();
-
 /**
- * Register a plugin extension that other plugins can access
- * @param pluginName - Name of the plugin registering the extension
- * @param extensionName - Name of the extension (e.g., "p2p", "discord")
- * @param extension - The extension object with methods/properties
+ * Plugin system - thin re-export barrel.
+ *
+ * All implementation has been extracted into focused modules under src/plugins/.
+ * This file preserves the original public API so existing imports keep working.
  */
-export function registerPluginExtension(pluginName: string, extensionName: string, extension: unknown): void {
-  const key = `${pluginName}.${extensionName}`;
-  pluginExtensions.set(key, extension);
-  logger.info(`[plugins] Extension registered: ${key}`);
-}
 
-/**
- * Unregister a plugin extension
- */
-export function unregisterPluginExtension(pluginName: string, extensionName: string): void {
-  const key = `${pluginName}.${extensionName}`;
-  pluginExtensions.delete(key);
-  logger.info(`[plugins] Extension unregistered: ${key}`);
-}
+// Extensions
+export { getPluginExtension, listPluginExtensions, registerPluginExtension, unregisterPluginExtension } from "./plugins/extensions.js";
 
-/**
- * Get a plugin extension by name
- * @param extensionName - Full name (plugin.extension) or just extension name
- */
-export function getPluginExtension<T = unknown>(extensionName: string): T | undefined {
-  // Try exact match first
-  if (pluginExtensions.has(extensionName)) {
-    return pluginExtensions.get(extensionName) as T;
-  }
-  // Try to find by extension name suffix
-  for (const [key, ext] of pluginExtensions) {
-    if (key.endsWith(`.${extensionName}`)) {
-      return ext as T;
-    }
-  }
-  return undefined;
-}
+// Installation
+export { disablePlugin, enablePlugin, getInstalledPlugins, installPlugin, listPlugins, removePlugin, uninstallPlugin } from "./plugins/installation.js";
+export type { InstallResult } from "./plugins/installation.js";
 
-/**
- * List all registered extensions
- */
-export function listPluginExtensions(): string[] {
-  return Array.from(pluginExtensions.keys());
-}
+// Loading
+export { getLoadedPlugin, loadAllPlugins, loadPlugin, shutdownAllPlugins, unloadPlugin } from "./plugins/loading.js";
+export type { LoadPluginOptions } from "./plugins/loading.js";
 
-function channelKey(channel: ChannelRef): string {
-  return `${channel.type}:${channel.id}`;
-}
+// Config schemas
+export { getConfigSchemas, listConfigSchemas } from "./plugins/config-schemas.js";
 
-// ============================================================================
-// JSON Schema to Zod Converter (for A2A tool registration)
-// ============================================================================
+// Registry & discovery
+export { addRegistry, discoverVoicePlugins, getPluginRegistries, listRegistries, removeRegistry, searchPlugins } from "./plugins/registry.js";
+export type { DiscoveredPlugin } from "./plugins/registry.js";
 
-/**
- * Convert a JSON Schema object to a Zod schema
- * Handles basic types: string, number, boolean, array, object
- */
-function jsonSchemaToZod(schema: Record<string, unknown>): z.ZodTypeAny {
-  if (!schema || typeof schema !== "object") {
-    return z.any();
-  }
-
-  const type = schema.type as string;
-  const properties = schema.properties as Record<string, Record<string, unknown>> | undefined;
-  const required = (schema.required as string[]) || [];
-  const items = schema.items as Record<string, unknown> | undefined;
-  const description = schema.description as string | undefined;
-
-  let zodSchema: z.ZodTypeAny;
-
-  switch (type) {
-    case "string":
-      zodSchema = z.string();
-      break;
-    case "number":
-    case "integer":
-      zodSchema = z.number();
-      break;
-    case "boolean":
-      zodSchema = z.boolean();
-      break;
-    case "array":
-      zodSchema = items ? z.array(jsonSchemaToZod(items)) : z.array(z.any());
-      break;
-    case "object":
-      if (properties) {
-        const shape: Record<string, z.ZodTypeAny> = {};
-        for (const [key, propSchema] of Object.entries(properties)) {
-          let prop = jsonSchemaToZod(propSchema);
-          // Add description if present
-          const propDesc = propSchema.description as string | undefined;
-          if (propDesc) {
-            prop = prop.describe(propDesc);
-          }
-          // Make optional if not in required array
-          if (!required.includes(key)) {
-            prop = prop.optional();
-          }
-          shape[key] = prop;
-        }
-        zodSchema = z.object(shape);
-      } else {
-        zodSchema = z.record(z.string(), z.any());
-      }
-      break;
-    default:
-      zodSchema = z.any();
-  }
-
-  if (description) {
-    zodSchema = zodSchema.describe(description);
-  }
-
-  return zodSchema;
-}
-
-/**
- * Register an A2A server with multiple tools
- * Converts A2AToolDefinition to RegisteredTool format
- */
-function registerA2AServerImpl(config: A2AServerConfig): void {
-  logger.info(`[plugins] Registering A2A server: ${config.name} (${config.tools.length} tools)`);
-
-  for (const tool of config.tools) {
-    try {
-      // Convert JSON Schema to Zod
-      const zodSchema = jsonSchemaToZod(tool.inputSchema) as z.ZodObject<any>;
-
-      // Register with a2a-mcp
-      registerA2ATool({
-        name: tool.name,
-        description: tool.description,
-        schema: zodSchema,
-        handler: async (args, _context) => {
-          // Call the plugin's handler
-          const result = await tool.handler(args);
-          return result;
-        },
-      });
-
-      logger.info(`[plugins]   ✓ Registered A2A tool: ${tool.name}`);
-    } catch (err) {
-      logger.error(`[plugins]   ✗ Failed to register A2A tool ${tool.name}:`, err);
-    }
-  }
-}
-
-// ============================================================================
-// Plugin Installation
-// ============================================================================
-
-export interface InstallResult {
-  name: string;
-  version: string;
-  path: string;
-  enabled: boolean;
-}
-
-// Validate input for safe shell interpolation
-const SAFE_NAME = /^[a-zA-Z0-9._-]+$/;
-const SAFE_PKG = /^[@a-zA-Z0-9._/-]+$/;
-
-function assertSafeName(value: string, label: string): void {
-  if (!SAFE_NAME.test(value)) {
-    throw new Error(`Invalid ${label}: ${value}`);
-  }
-}
-
-export async function installPlugin(source: string): Promise<InstalledPlugin> {
-  mkdirSync(PLUGINS_DIR, { recursive: true });
-
-  // Determine source type
-  if (source.startsWith("github:")) {
-    // GitHub repo
-    const repo = source.replace("github:", "");
-    const parts = repo.split("/");
-    if (parts.length !== 2) throw new Error("GitHub source must be github:owner/repo");
-    assertSafeName(parts[0], "GitHub owner");
-    assertSafeName(parts[1], "GitHub repo");
-    const pluginDir = join(PLUGINS_DIR, parts[1]);
-
-    // Clone or pull
-    if (existsSync(pluginDir)) {
-      execFileSync("git", ["pull"], { cwd: pluginDir, stdio: "inherit" });
-    } else {
-      execFileSync("git", ["clone", `https://github.com/${parts[0]}/${parts[1]}`, pluginDir], { stdio: "inherit" });
-    }
-
-    // Install dependencies if package.json exists
-    const pkgPath = join(pluginDir, "package.json");
-    if (existsSync(pkgPath)) {
-      logger.info(`[plugins] Installing dependencies for ${repo}...`);
-      execFileSync("npm", ["install"], { cwd: pluginDir, stdio: "inherit" });
-
-      // Build TypeScript plugins if tsconfig.json exists
-      if (existsSync(join(pluginDir, "tsconfig.json"))) {
-        logger.info(`[plugins] Building TypeScript plugin...`);
-        execFileSync("npm", ["run", "build"], { cwd: pluginDir, stdio: "inherit" });
-      }
-    }
-
-    // Read package.json for metadata
-    const pkg = existsSync(pkgPath) ? JSON.parse(readFileSync(pkgPath, "utf-8")) : {};
-
-    const installed: InstalledPlugin = {
-      name: pkg.name || repo.split("/")[1] || repo,
-      version: pkg.version || "0.0.0",
-      description: pkg.description,
-      source: "github",
-      path: pluginDir,
-      enabled: false,
-      installedAt: Date.now(),
-    };
-
-    addInstalledPlugin(installed);
-    return installed;
-  } else if (source.startsWith("./") || source.startsWith("/") || source.startsWith("~/")) {
-    // Local path
-    const resolved = resolve(source.replace("~", process.env.HOME || "~"));
-    const pluginDir = join(PLUGINS_DIR, resolved.split("/").pop() || "plugin");
-
-    // Symlink (no shell — avoids injection via path)
-    if (!existsSync(pluginDir)) {
-      symlinkSync(resolved, pluginDir);
-    }
-
-    // Install dependencies if package.json exists
-    const pkgPath = join(pluginDir, "package.json");
-    if (existsSync(pkgPath)) {
-      logger.info(`[plugins] Installing dependencies for local plugin...`);
-      execFileSync("npm", ["install"], { cwd: pluginDir, stdio: "inherit" });
-
-      // Build TypeScript plugins if tsconfig.json exists
-      if (existsSync(join(pluginDir, "tsconfig.json"))) {
-        logger.info(`[plugins] Building TypeScript plugin...`);
-        execFileSync("npm", ["run", "build"], { cwd: pluginDir, stdio: "inherit" });
-      }
-    }
-
-    const pkg = existsSync(pkgPath) ? JSON.parse(readFileSync(pkgPath, "utf-8")) : {};
-
-    const installed: InstalledPlugin = {
-      name: pkg.name || resolved.split("/").pop() || "plugin",
-      version: pkg.version || "0.0.0",
-      description: pkg.description,
-      source: "local",
-      path: pluginDir,
-      enabled: false,
-      installedAt: Date.now(),
-    };
-
-    addInstalledPlugin(installed);
-    return installed;
-  } else {
-    // npm package - normalize to wopr-plugin-<name> format (accept wopr-<name> too)
-    const shortName = source.replace(/^wopr-plugin-/, "").replace(/^wopr-/, "");
-    const npmPackage =
-      source.startsWith("wopr-") && !source.startsWith("wopr-plugin-") ? source : `wopr-plugin-${shortName}`;
-    if (!SAFE_PKG.test(npmPackage)) throw new Error(`Invalid npm package name: ${npmPackage}`);
-    const pluginDir = join(PLUGINS_DIR, shortName);
-    mkdirSync(pluginDir, { recursive: true });
-
-    // Use npm to install
-    execFileSync("npm", ["install", npmPackage], { cwd: pluginDir, stdio: "inherit" });
-
-    // Read installed package metadata
-    const pkgPath = join(pluginDir, "node_modules", npmPackage, "package.json");
-    const pkg = existsSync(pkgPath) ? JSON.parse(readFileSync(pkgPath, "utf-8")) : {};
-
-    const installed: InstalledPlugin = {
-      name: shortName,
-      version: pkg.version || "0.0.0",
-      description: pkg.description,
-      source: "npm",
-      path: join(pluginDir, "node_modules", npmPackage),
-      enabled: false,
-      installedAt: Date.now(),
-    };
-
-    addInstalledPlugin(installed);
-    return installed;
-  }
-}
-
-export function removePlugin(name: string): boolean {
-  return uninstallPlugin(name);
-}
-
-export function uninstallPlugin(name: string): boolean {
-  const installed = getInstalledPlugins();
-  const plugin = installed.find((p) => p.name === name);
-  if (!plugin) return false;
-
-  // Remove files (only if under PLUGINS_DIR to prevent path traversal)
-  const normalizedPath = resolve(plugin.path);
-  const normalizedBase = resolve(PLUGINS_DIR);
-  if (existsSync(normalizedPath) && normalizedPath.startsWith(`${normalizedBase}/`)) {
-    rmSync(normalizedPath, { recursive: true, force: true });
-  }
-
-  // Remove from registry
-  const remaining = installed.filter((p) => p.name !== name);
-  writeFileSync(PLUGINS_FILE, JSON.stringify(remaining, null, 2));
-
-  return true;
-}
-
-export function enablePlugin(name: string): boolean {
-  const installed = getInstalledPlugins();
-  const plugin = installed.find((p) => p.name === name);
-  if (!plugin) return false;
-
-  plugin.enabled = true;
-  writeFileSync(PLUGINS_FILE, JSON.stringify(installed, null, 2));
-  return true;
-}
-
-export function disablePlugin(name: string): boolean {
-  const installed = getInstalledPlugins();
-  const plugin = installed.find((p) => p.name === name);
-  if (!plugin) return false;
-
-  plugin.enabled = false;
-  writeFileSync(PLUGINS_FILE, JSON.stringify(installed, null, 2));
-  return true;
-}
-
-export function listPlugins(): InstalledPlugin[] {
-  return getInstalledPlugins();
-}
-
-export function getInstalledPlugins(): InstalledPlugin[] {
-  if (!existsSync(PLUGINS_FILE)) return [];
-  return JSON.parse(readFileSync(PLUGINS_FILE, "utf-8"));
-}
-
-function addInstalledPlugin(plugin: InstalledPlugin): void {
-  const installed = getInstalledPlugins();
-  const existing = installed.findIndex((p) => p.name === plugin.name);
-  if (existing >= 0) {
-    installed[existing] = plugin;
-  } else {
-    installed.push(plugin);
-  }
-  writeFileSync(PLUGINS_FILE, JSON.stringify(installed, null, 2));
-}
-
-// ============================================================================
-// Plugin Context Creation
-// ============================================================================
-
-function createPluginContext(
-  plugin: InstalledPlugin,
-  injectors: {
-    inject: (session: string, message: string, options?: PluginInjectOptions) => Promise<string>;
-    getSessions: () => string[];
-  },
-): WOPRPluginContext {
-  const pluginName = plugin.name;
-
-  return {
-    inject: (session: string, message: string, options?: PluginInjectOptions) =>
-      injectors.inject(session, message, options),
-    logMessage: (
-      session: string,
-      message: string,
-      options?: { from?: string; senderId?: string; channel?: ChannelRef },
-    ) => logMessageToSession(session, message, options),
-    getSessions: injectors.getSessions,
-    cancelInject: (session: string) => cancelSessionInject(session),
-
-    // Event bus - reactive primitive for plugin composition
-    events: createPluginEventBus(pluginName),
-
-    // Hooks - typed shorthand for common event patterns
-    hooks: createPluginHookManager(pluginName),
-
-    registerContextProvider(provider: ContextProvider) {
-      registerCtxProvider(provider);
-    },
-
-    unregisterContextProvider(name: string) {
-      unregisterCtxProvider(name);
-    },
-
-    getContextProvider(name: string): ContextProvider | undefined {
-      return getCtxProvider(name);
-    },
-
-    registerChannel(adapter: ChannelAdapter) {
-      channelAdapters.set(channelKey(adapter.channel), adapter);
-    },
-
-    unregisterChannel(channel: ChannelRef) {
-      channelAdapters.delete(channelKey(channel));
-    },
-
-    getChannel(channel: ChannelRef) {
-      return channelAdapters.get(channelKey(channel));
-    },
-
-    getChannels() {
-      return Array.from(channelAdapters.values());
-    },
-
-    getChannelsForSession(session: string) {
-      return Array.from(channelAdapters.values()).filter((adapter) => adapter.session === session);
-    },
-
-    registerWebUiExtension(extension: WebUiExtension) {
-      webUiExtensions.set(`${pluginName}:${extension.id}`, extension);
-    },
-
-    unregisterWebUiExtension(id: string) {
-      webUiExtensions.delete(`${pluginName}:${id}`);
-    },
-
-    getWebUiExtensions() {
-      return Array.from(webUiExtensions.values());
-    },
-
-    registerUiComponent(extension: UiComponentExtension) {
-      uiComponents.set(`${pluginName}:${extension.id}`, extension);
-    },
-
-    unregisterUiComponent(id: string) {
-      uiComponents.delete(`${pluginName}:${id}`);
-    },
-
-    getUiComponents() {
-      return Array.from(uiComponents.values());
-    },
-
-    getConfig<T>(): T {
-      // Load from central config
-      const cfg = centralConfig.get();
-      return (cfg.plugins.data?.[pluginName] || {}) as T;
-    },
-
-    async saveConfig<T>(pluginConfig: T): Promise<void> {
-      // Save to central config
-      await centralConfig.load();
-      const cfg = centralConfig.get();
-      if (!cfg.plugins.data) cfg.plugins.data = {};
-      cfg.plugins.data[pluginName] = pluginConfig;
-      centralConfig.setValue("plugins.data", cfg.plugins.data);
-      await centralConfig.save();
-    },
-
-    getMainConfig(key?: string): any {
-      // Access main WOPR config (read-only)
-      const cfg = centralConfig.get();
-      if (!key) return cfg;
-      return centralConfig.getValue(key);
-    },
-
-    registerProvider(provider: ModelProvider) {
-      logger.info(`[plugins] Provider registration: ${provider.id} (${provider.name})`);
-      providerPlugins.set(provider.id, provider);
-      providerRegistry.register(provider);
-      logger.info(`[plugins]   ✓ Provider ${provider.id} registered in registry`);
-    },
-
-    unregisterProvider(id: string) {
-      providerPlugins.delete(id);
-      // Note: providerRegistry doesn't have unregister, providers are removed from runtime only
-    },
-
-    getProvider(id: string): ModelProvider | undefined {
-      return (
-        providerPlugins.get(id) ||
-        (providerRegistry.listProviders().find((p) => p.id === id) as unknown as ModelProvider)
-      );
-    },
-
-    async getAgentIdentity() {
-      // Resolve from workspace IDENTITY.md
-      return await resolveIdentity();
-    },
-
-    async getUserProfile() {
-      // Resolve from workspace USER.md
-      return await resolveUserProfile();
-    },
-
-    registerConfigSchema(pluginId: string, schema: ConfigSchema) {
-      configSchemas.set(pluginId, schema);
-    },
-
-    unregisterConfigSchema(pluginId: string) {
-      configSchemas.delete(pluginId);
-    },
-
-    getConfigSchema(pluginId: string): ConfigSchema | undefined {
-      return configSchemas.get(pluginId);
-    },
-
-    // Plugin extensions - allow plugins to expose APIs to other plugins
-    registerExtension(name: string, extension: unknown) {
-      registerPluginExtension(pluginName, name, extension);
-    },
-
-    unregisterExtension(name: string) {
-      unregisterPluginExtension(pluginName, name);
-    },
-
-    getExtension<T = unknown>(name: string): T | undefined {
-      return getPluginExtension<T>(name);
-    },
-
-    listExtensions(): string[] {
-      return listPluginExtensions();
-    },
-
-    // Voice extensions - allow voice plugins to register STT/TTS providers
-    // and channel plugins to discover and use them
-    registerSTTProvider(provider: STTProvider) {
-      const voiceRegistry = getVoiceRegistry();
-      voiceRegistry.registerSTT(provider);
-      // Also register as an extension for discovery via getExtension('stt')
-      registerPluginExtension(pluginName, "stt", provider);
-      logger.info(`[plugins] STT provider registered: ${provider.metadata.name}`);
-    },
-
-    registerTTSProvider(provider: TTSProvider) {
-      const voiceRegistry = getVoiceRegistry();
-      voiceRegistry.registerTTS(provider);
-      // Also register as an extension for discovery via getExtension('tts')
-      registerPluginExtension(pluginName, "tts", provider);
-      logger.info(`[plugins] TTS provider registered: ${provider.metadata.name}`);
-    },
-
-    getSTT(): STTProvider | null {
-      return getVoiceRegistry().getSTT();
-    },
-
-    getTTS(): TTSProvider | null {
-      return getVoiceRegistry().getTTS();
-    },
-
-    hasVoice(): { stt: boolean; tts: boolean } {
-      const voiceRegistry = getVoiceRegistry();
-      return {
-        stt: voiceRegistry.getSTT() !== null,
-        tts: voiceRegistry.getTTS() !== null,
-      };
-    },
-
-    // Channel providers - channel plugins register themselves for cross-plugin protocol commands
-    registerChannelProvider(provider: ChannelProvider) {
-      registerChannelProviderCore(provider);
-    },
-
-    unregisterChannelProvider(id: string) {
-      unregisterChannelProviderCore(id);
-    },
-
-    getChannelProvider(id: string): ChannelProvider | undefined {
-      return getChannelProviderCore(id);
-    },
-
-    getChannelProviders(): ChannelProvider[] {
-      return getChannelProvidersCore();
-    },
-
-    // A2A tools - plugins register MCP tools for agent-to-agent communication
-    registerA2AServer(config: A2AServerConfig) {
-      registerA2AServerImpl(config);
-    },
-
-    log: createPluginLogger(plugin.name),
-
-    getPluginDir(): string {
-      return plugin.source === "local" ? plugin.path : join(PLUGINS_DIR, plugin.name);
-    },
-  };
-}
-
-/**
- * Create an event bus instance scoped to a plugin
- */
-function createPluginEventBus(pluginName: string): WOPREventBus {
-  // Track original -> wrapped handler mapping so off() can find the right listener
-  const handlerMap = new Map<EventHandler<any>, EventHandler<any>>();
-
-  return {
-    on<T extends keyof import("./types.js").WOPREventMap>(
-      event: T,
-      handler: EventHandler<import("./types.js").WOPREventMap[T]>,
-    ): () => void {
-      // Wrap handler to identify plugin source
-      const wrappedHandler: EventHandler<any> = async (payload, evt) => {
-        // Add plugin source to event metadata
-        const eventWithSource = { ...evt, source: pluginName };
-        await handler(payload, eventWithSource);
-      };
-
-      handlerMap.set(handler, wrappedHandler);
-      return eventBus.on(event, wrappedHandler);
-    },
-
-    once<T extends keyof import("./types.js").WOPREventMap>(
-      event: T,
-      handler: EventHandler<import("./types.js").WOPREventMap[T]>,
-    ): void {
-      const wrappedHandler: EventHandler<any> = async (payload, evt) => {
-        const eventWithSource = { ...evt, source: pluginName };
-        await handler(payload, eventWithSource);
-      };
-      handlerMap.set(handler, wrappedHandler);
-      eventBus.once(event, wrappedHandler);
-    },
-
-    off<T extends keyof import("./types.js").WOPREventMap>(
-      event: T,
-      handler: EventHandler<import("./types.js").WOPREventMap[T]>,
-    ): void {
-      const wrapped = handlerMap.get(handler);
-      if (wrapped) {
-        eventBus.off(event, wrapped as EventHandler<import("./types.js").WOPREventMap[T]>);
-        handlerMap.delete(handler);
-      }
-    },
-
-    async emit(event: string, payload: any): Promise<void> {
-      await eventBus.emit(event as any, payload, pluginName);
-    },
-
-    async emitCustom(event: string, payload: any): Promise<void> {
-      await eventBus.emitCustom(event, payload, pluginName);
-    },
-
-    listenerCount(event: string): number {
-      return eventBus.listenerCount(event);
-    },
-  };
-}
-
-/**
- * Hook registration entry with metadata
- */
-interface HookEntry {
-  handler: (...args: any[]) => any;
-  priority: number;
-  name?: string;
-  once: boolean;
-  unsubscribe: () => void;
-}
-
-/**
- * Create a hook manager scoped to a plugin
- * Hooks provide typed, mutable access to core lifecycle events
- * with priority ordering (lower = runs first)
- */
-function createPluginHookManager(_pluginName: string): WOPRHookManager {
-  // Map of event -> array of hook entries (sorted by priority)
-  const hookEntries = new Map<string, HookEntry[]>();
-
-  // One bus subscription per event (prevents N*N handler calls)
-  const busSubscriptions = new Map<string, () => void>();
-
-  // Mutable events that can transform data or block
-  const mutableEvents = new Set(["message:incoming", "message:outgoing", "channel:message"]);
-
-  // Map hook event names to underlying event bus events
-  const eventMapping: Record<string, string> = {
-    "message:incoming": "session:beforeInject",
-    "message:outgoing": "session:afterInject",
-  };
-
-  function getEntries(event: string): HookEntry[] {
-    if (!hookEntries.has(event)) {
-      hookEntries.set(event, []);
-    }
-    return hookEntries.get(event)!;
-  }
-
-  function insertSorted(entries: HookEntry[], entry: HookEntry): void {
-    // Insert in priority order (lower = first)
-    const idx = entries.findIndex((e) => e.priority > entry.priority);
-    if (idx === -1) {
-      entries.push(entry);
-    } else {
-      entries.splice(idx, 0, entry);
-    }
-  }
-
-  /** Ensure exactly one event bus listener exists for this hook event */
-  function ensureBusSubscription(event: string): void {
-    if (busSubscriptions.has(event)) return;
-
-    const busEvent = eventMapping[event] || event;
-    const isMutable = mutableEvents.has(event);
-
-    const unsubscribe = eventBus.on(busEvent as any, async (payload, _evt) => {
-      const entries = getEntries(event);
-
-      if (isMutable) {
-        let prevented = false;
-        const mutableEvent: MutableHookEvent<any> = {
-          data: payload,
-          session: payload.session || "default",
-          preventDefault() {
-            prevented = true;
-            if (payload && typeof payload === "object") {
-              (payload as any)._prevented = true;
-            }
-          },
-          isPrevented() {
-            return prevented;
-          },
-        };
-
-        for (const entry of [...entries]) {
-          await entry.handler(mutableEvent);
-
-          if (entry.once) {
-            const idx = entries.indexOf(entry);
-            if (idx !== -1) entries.splice(idx, 1);
-          }
-
-          if (mutableEvent.isPrevented()) break;
-        }
-      } else {
-        for (const entry of [...entries]) {
-          await entry.handler(payload);
-
-          if (entry.once) {
-            const idx = entries.indexOf(entry);
-            if (idx !== -1) entries.splice(idx, 1);
-          }
-        }
-      }
-
-      // If all entries removed, clean up bus subscription
-      if (entries.length === 0) {
-        unsubscribe();
-        busSubscriptions.delete(event);
-      }
-    });
-
-    busSubscriptions.set(event, unsubscribe);
-  }
-
-  return {
-    on(event: string, handler: (...args: any[]) => any, options?: HookOptions): () => void {
-      const priority = options?.priority ?? 100;
-      const name = options?.name;
-      const once = options?.once ?? false;
-
-      const entry: HookEntry = {
-        handler,
-        priority,
-        name,
-        once,
-        unsubscribe: () => {}, // placeholder, removal handled below
-      };
-
-      const entries = getEntries(event);
-      insertSorted(entries, entry);
-
-      // Subscribe to event bus (only once per event)
-      ensureBusSubscription(event);
-
-      return () => {
-        const entries = getEntries(event);
-        const idx = entries.indexOf(entry);
-        if (idx !== -1) entries.splice(idx, 1);
-
-        // If no entries left, unsubscribe from bus
-        if (entries.length === 0) {
-          busSubscriptions.get(event)?.();
-          busSubscriptions.delete(event);
-        }
-      };
-    },
-
-    off(event: string, handler: (...args: any[]) => any): void {
-      const entries = getEntries(event);
-      const idx = entries.findIndex((e) => e.handler === handler);
-      if (idx !== -1) {
-        entries.splice(idx, 1);
-      }
-      // If no entries left, unsubscribe from bus
-      if (entries.length === 0) {
-        busSubscriptions.get(event)?.();
-        busSubscriptions.delete(event);
-      }
-    },
-
-    offByName(name: string): void {
-      for (const [event, entries] of hookEntries) {
-        const toRemove = entries.filter((e) => e.name === name);
-        for (const entry of toRemove) {
-          const idx = entries.indexOf(entry);
-          if (idx !== -1) entries.splice(idx, 1);
-        }
-        // If no entries left, unsubscribe from bus
-        if (entries.length === 0) {
-          busSubscriptions.get(event)?.();
-          busSubscriptions.delete(event);
-        }
-      }
-    },
-
-    list(): Array<{ event: string; name?: string; priority: number }> {
-      const result: Array<{ event: string; name?: string; priority: number }> = [];
-      for (const [event, entries] of hookEntries) {
-        for (const entry of entries) {
-          result.push({
-            event,
-            name: entry.name,
-            priority: entry.priority,
-          });
-        }
-      }
-      return result.sort((a, b) => a.priority - b.priority);
-    },
-  } as WOPRHookManager;
-}
-
-/** Options for loading plugins */
-export interface LoadPluginOptions {
-  /** Automatically install missing dependencies */
-  autoInstall?: boolean;
-  /** Skip requirements check entirely */
-  skipRequirementsCheck?: boolean;
-  /** Prompt function for interactive install */
-  promptInstall?: (message: string) => Promise<boolean>;
-  /** Skip plugin init (for CLI commands that just need the plugin module) */
-  skipInit?: boolean;
-}
-
-export async function loadPlugin(
-  installed: InstalledPlugin,
-  injectors: {
-    inject: (session: string, message: string, options?: PluginInjectOptions) => Promise<string>;
-    getSessions: () => string[];
-  },
-  options: LoadPluginOptions = {},
-): Promise<WOPRPlugin> {
-  // Find the entry point and read package.json
-  let entryPoint = installed.path;
-  let pkg: any = {};
-
-  if (existsSync(join(installed.path, "package.json"))) {
-    pkg = JSON.parse(readFileSync(join(installed.path, "package.json"), "utf-8"));
-    entryPoint = join(installed.path, pkg.main || "index.js");
-  } else if (existsSync(join(installed.path, "index.js"))) {
-    entryPoint = join(installed.path, "index.js");
-  } else if (existsSync(join(installed.path, "index.ts"))) {
-    entryPoint = join(installed.path, "index.ts");
-  }
-
-  // Check requirements from package.json wopr.plugin metadata
-  if (!options.skipRequirementsCheck) {
-    const pluginMeta = pkg.wopr?.plugin;
-    const requires: VoicePluginRequirements | undefined = pluginMeta?.requires;
-    const installMethods: InstallMethod[] | undefined = pluginMeta?.install;
-
-    if (requires) {
-      logger.info(`[plugins] Checking requirements for ${installed.name}...`);
-
-      const {
-        satisfied,
-        installed: installedDeps,
-        errors,
-      } = await ensureRequirements(requires, installMethods, {
-        auto: options.autoInstall,
-        prompt: options.promptInstall,
-      });
-
-      if (!satisfied) {
-        const check = await checkRequirements(requires);
-        const missing = formatMissingRequirements(check);
-        const errorDetail = errors.length > 0 ? `\nInstall errors:\n${errors.map((e) => `  - ${e}`).join("\n")}` : "";
-        throw new Error(`Plugin ${installed.name} requirements not satisfied:\n${missing}${errorDetail}`);
-      }
-
-      if (installedDeps.length > 0) {
-        logger.info(`[plugins] Installed ${installedDeps.length} dependencies for ${installed.name}`);
-      }
-    }
-  }
-
-  // Temporarily change cwd to plugin directory for proper module resolution
-  const originalCwd = process.cwd();
-  process.chdir(installed.path);
-
-  let module: any;
-  try {
-    // Dynamic import with cache-busting query param for reloads
-    // ESM caches by URL, so adding timestamp forces fresh import
-    const cacheBuster = `?t=${Date.now()}`;
-    module = await import(entryPoint + cacheBuster);
-  } finally {
-    process.chdir(originalCwd);
-  }
-  const plugin: WOPRPlugin = module.default || module;
-
-  // Create context
-  const context = createPluginContext(installed, injectors);
-
-  // Store
-  loadedPlugins.set(installed.name, { plugin, context });
-
-  // Initialize if needed (skip for CLI commands)
-  if (plugin.init && !options.skipInit) {
-    await plugin.init(context);
-  }
-
-  return plugin;
-}
-
-export async function unloadPlugin(name: string): Promise<void> {
-  const loaded = loadedPlugins.get(name);
-  if (!loaded) return;
-
-  // Shutdown if needed
-  if (loaded.plugin.shutdown) {
-    await loaded.plugin.shutdown();
-  }
-
-  // Clean up registrations
-  if (loaded.plugin.commands) {
-    // Commands are registered per-plugin, no global registry to clean
-  }
-
-  loadedPlugins.delete(name);
-}
-
-export function getLoadedPlugin(name: string): { plugin: WOPRPlugin; context: WOPRPluginContext } | undefined {
-  return loadedPlugins.get(name);
-}
-
-export function getWebUiExtensions(): WebUiExtension[] {
-  return Array.from(webUiExtensions.values());
-}
-
-export function getUiComponents(): UiComponentExtension[] {
-  return Array.from(uiComponents.values());
-}
-
-// ============================================================================
-// Config Schemas
-// ============================================================================
-
-export function getConfigSchemas(): Map<string, ConfigSchema> {
-  return configSchemas;
-}
-
-export function listConfigSchemas(): { pluginId: string; schema: ConfigSchema }[] {
-  return Array.from(configSchemas.entries()).map(([pluginId, schema]) => ({
-    pluginId,
-    schema,
-  }));
-}
-
-// ============================================================================
-// Plugin Registry
-// ============================================================================
-
-export function getPluginRegistries(): PluginRegistryEntry[] {
-  if (!existsSync(REGISTRIES_FILE)) return [];
-  return JSON.parse(readFileSync(REGISTRIES_FILE, "utf-8"));
-}
-
-export function addRegistry(url: string, name?: string): PluginRegistryEntry {
-  const registries = getPluginRegistries();
-  const entry: PluginRegistryEntry = {
-    url,
-    name: name || new URL(url).hostname,
-    enabled: true,
-    lastSync: 0,
-  };
-  registries.push(entry);
-  writeFileSync(REGISTRIES_FILE, JSON.stringify(registries, null, 2));
-  return entry;
-}
-
-export function removeRegistry(url: string): boolean {
-  const registries = getPluginRegistries();
-  const filtered = registries.filter((r) => r.url !== url);
-  if (filtered.length === registries.length) return false;
-  writeFileSync(REGISTRIES_FILE, JSON.stringify(filtered, null, 2));
-  return true;
-}
-
-export function listRegistries(): PluginRegistryEntry[] {
-  return getPluginRegistries();
-}
-
-export interface DiscoveredPlugin {
-  name: string;
-  description?: string;
-  source: "github" | "npm" | "installed" | "registry";
-  url?: string;
-  version?: string;
-  installed?: boolean;
-}
-
-/**
- * Search for plugins across multiple sources
- */
-export async function searchPlugins(query: string): Promise<DiscoveredPlugin[]> {
-  const results: DiscoveredPlugin[] = [];
-  const seen = new Set<string>();
-
-  // 1. Check installed plugins first
-  const installed = getInstalledPlugins();
-  for (const p of installed) {
-    if (!query || p.name.includes(query)) {
-      if (!seen.has(p.name)) {
-        seen.add(p.name);
-        results.push({
-          name: p.name,
-          description: p.description,
-          source: "installed",
-          version: p.version,
-          installed: true,
-        });
-      }
-    }
-  }
-
-  // 2. Search GitHub repos (if gh is available)
-  try {
-    const ghResults = await searchGitHubPlugins(query);
-    for (const p of ghResults) {
-      if (!seen.has(p.name)) {
-        seen.add(p.name);
-        p.installed = installed.some((i) => i.name === p.name);
-        results.push(p);
-      }
-    }
-  } catch {
-    // gh not available or error, skip
-  }
-
-  // 3. Search npm (if online)
-  try {
-    const npmResults = await searchNpmPlugins(query);
-    for (const p of npmResults) {
-      if (!seen.has(p.name)) {
-        seen.add(p.name);
-        p.installed = installed.some((i) => i.name === p.name);
-        results.push(p);
-      }
-    }
-  } catch {
-    // npm search failed, skip
-  }
-
-  return results;
-}
-
-/**
- * Search GitHub for wopr plugins using gh CLI
- */
-async function searchGitHubPlugins(query?: string): Promise<DiscoveredPlugin[]> {
-  const results: DiscoveredPlugin[] = [];
-
-  try {
-    // Get user's repos matching wopr-plugin-*
-    const output = execFileSync("gh", ["repo", "list", "--json", "name,description,url", "--limit", "100"], {
-      encoding: "utf-8",
-      timeout: 10000,
-    });
-    const repos = JSON.parse(output);
-
-    for (const repo of repos) {
-      if (repo.name.startsWith("wopr-plugin-")) {
-        if (!query || repo.name.includes(query) || repo.description?.includes(query)) {
-          results.push({
-            name: repo.name,
-            description: repo.description,
-            source: "github",
-            url: repo.url,
-          });
-        }
-      }
-    }
-  } catch {
-    // gh not available or not authenticated
-  }
-
-  return results;
-}
-
-/**
- * Search npm for wopr plugins
- */
-async function searchNpmPlugins(query?: string): Promise<DiscoveredPlugin[]> {
-  const results: DiscoveredPlugin[] = [];
-  const sanitized = query ? query.replace(/[^a-zA-Z0-9._-]/g, "") : "";
-  const searchTerm = sanitized ? `wopr-plugin-${sanitized}` : "wopr-plugin-";
-
-  try {
-    const result = spawnSync("npm", ["search", searchTerm, "--json"], {
-      encoding: "utf-8",
-      timeout: 15000,
-      maxBuffer: 1024 * 1024,
-      shell: false,
-    });
-    if (result.error || result.status !== 0 || !result.stdout) {
-      return results;
-    }
-    const output = result.stdout;
-    const packages = JSON.parse(output);
-
-    for (const pkg of packages) {
-      if (pkg.name.startsWith("wopr-plugin-")) {
-        results.push({
-          name: pkg.name,
-          description: pkg.description,
-          source: "npm",
-          version: pkg.version,
-          url: `https://www.npmjs.com/package/${pkg.name}`,
-        });
-      }
-    }
-  } catch {
-    // npm search failed
-  }
-
-  return results;
-}
-
-/**
- * Discover all available voice plugins from GitHub
- */
-export async function discoverVoicePlugins(): Promise<{
-  stt: DiscoveredPlugin[];
-  tts: DiscoveredPlugin[];
-  channels: DiscoveredPlugin[];
-  cli: DiscoveredPlugin[];
-}> {
-  const all = await searchPlugins("voice");
-
-  return {
-    stt: all.filter((p) => p.name.includes("stt") || p.name.includes("whisper") || p.name.includes("deepgram")),
-    tts: all.filter((p) => p.name.includes("tts") || p.name.includes("piper") || p.name.includes("elevenlabs")),
-    channels: all.filter((p) => p.name.includes("channel") && p.name.includes("voice")),
-    cli: all.filter((p) => p.name.includes("voice-cli")),
-  };
-}
-
-// ============================================================================
-// Context Providers
-// ============================================================================
-
-export function getContextProvider(session: string): ContextProvider | undefined {
-  return contextProviders.get(session);
-}
-
-export function getChannel(channel: ChannelRef): ChannelAdapter | undefined {
-  return channelAdapters.get(channelKey(channel));
-}
-
-export function getChannels(): ChannelAdapter[] {
-  return Array.from(channelAdapters.values());
-}
-
-export function getChannelsForSession(session: string): ChannelAdapter[] {
-  return Array.from(channelAdapters.values()).filter((adapter) => adapter.session === session);
-}
-
-// ============================================================================
-// Plugin Logger
-// ============================================================================
-
-function createPluginLogger(pluginName: string): PluginLogger {
-  return {
-    info: (message: string, ...args: any[]) => {
-      logger.info(`[${pluginName}] ${message}`, ...args);
-    },
-    warn: (message: string, ...args: any[]) => {
-      logger.warn(`[${pluginName}] ${message}`, ...args);
-    },
-    error: (message: string, ...args: any[]) => {
-      logger.error(`[${pluginName}] ${message}`, ...args);
-    },
-    debug: (message: string, ...args: any[]) => {
-      if (process.env.DEBUG) {
-        logger.debug(`[${pluginName}] ${message}`, ...args);
-      }
-    },
-  };
-}
-
-// ============================================================================
-// Batch Plugin Operations
-// ============================================================================
-
-export async function loadAllPlugins(
-  injectors: {
-    inject: (session: string, message: string, options?: PluginInjectOptions) => Promise<string>;
-    getSessions: () => string[];
-  },
-  options: LoadPluginOptions = {},
-): Promise<void> {
-  logger.info(`[plugins] loadAllPlugins starting...`);
-  logger.info(`[plugins] WOPR_HOME: ${process.env.WOPR_HOME || "not set"}`);
-  if (options.autoInstall) {
-    logger.info(`[plugins] Auto-install enabled`);
-  }
-
-  const installed = getInstalledPlugins();
-  logger.info(`[plugins] Found ${installed.length} installed plugins`);
-
-  for (const p of installed) {
-    logger.info(`[plugins]  - ${p.name}: enabled=${p.enabled}, path=${p.path}`);
-  }
-
-  let loadedCount = 0;
-  const failed: { name: string; error: string }[] = [];
-
-  for (const plugin of installed) {
-    logger.info(`[plugins] Processing ${plugin.name}...`);
-    if (!plugin.enabled) {
-      logger.info(`[plugins]   Skipping ${plugin.name} (disabled)`);
-      continue;
-    }
-
-    logger.info(`[plugins]   Loading ${plugin.name} from ${plugin.path}...`);
-    try {
-      await loadPlugin(plugin, injectors, options);
-      loadedCount++;
-      logger.info(`[plugins]   ✓ Loaded: ${plugin.name}`);
-    } catch (err: any) {
-      logger.error(`[plugins]   ✗ Failed to load ${plugin.name}:`, err.message);
-      if (err.stack) {
-        logger.error(`[plugins]     Stack:`, err.stack.substring(0, 200));
-      }
-      failed.push({ name: plugin.name, error: err.message });
-    }
-  }
-
-  logger.info(`[plugins] loadAllPlugins complete. Loaded ${loadedCount}/${installed.length} plugins`);
-
-  if (failed.length > 0) {
-    logger.warn(`[plugins] ${failed.length} plugins failed to load:`);
-    for (const f of failed) {
-      logger.warn(`[plugins]   - ${f.name}: ${f.error.split("\n")[0]}`);
-    }
-  }
-}
-
-export async function shutdownAllPlugins(): Promise<void> {
-  for (const [name] of loadedPlugins) {
-    try {
-      await unloadPlugin(name);
-      logger.info(`[plugins] Unloaded: ${name}`);
-    } catch (err) {
-      logger.error(`[plugins] Failed to unload ${name}:`, err);
-    }
-  }
-}
+// Public accessors for runtime state
+export { getChannel, getChannels, getChannelsForSession, getContextProvider, getUiComponents, getWebUiExtensions } from "./plugins/accessors.js";
